rename db utils, create a new utils file for the battlefield and create a filter for get objects from battlefield
